use uint64_t for time

+++ src/engine/animation.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <optional>
+#include <stdint.h>
 #include <string>
 #include <unordered_map>
 #include <vector>
@@ -16,12 +17,12 @@ namespace engine {
 
 	struct Animation {
 		AnimationID id;
-		float start;
-		float length;
+		uint64_t start;
+		uint64_t length;
 		bool repeats;
 
-		bool is_playing(float global_time) const;
-		float local_time(float global_time) const;
+		bool is_playing(uint64_t global_time) const;
+		float local_time(uint64_t global_time) const;
 	};
 
 	class AnimationSystem {
@@ -29,13 +30,13 @@ namespace engine {
 		const std::vector<Animation>& animations(AnimationKey key) const;
 		std::optional<Animation> most_recent_animation(AnimationKey key) const;
 
-		AnimationID start_animation(AnimationKey key, float length, float start_time);
-		AnimationID start_single_shot_animation(AnimationKey key, float length, float start_time);
+		AnimationID start_animation(AnimationKey key, uint64_t length, uint64_t start_time);
+		AnimationID start_single_shot_animation(AnimationKey key, uint64_t length, uint64_t start_time);
 		void stop_animation(AnimationID id);
-		void clear_old_animations(float global_time);
+		void clear_old_animations(uint64_t global_time);
 
 	private:
-		AnimationID _start_animation(AnimationKey key, float length, float start_time, bool repeats);
+		AnimationID _start_animation(AnimationKey key, uint64_t length, uint64_t start_time, bool repeats);
 
 		const std::vector<Animation> m_empty_animation;
 		std::unordered_map<AnimationKey, std::vector<Animation>> m_animations;
